requirements.txt
yfinance
ccxt
psycopg2-binary